requirements.txt
fastapi
uvicorn[standard]
python-dotenv
twilio
huggingface_hub
langchain
langchain-community
ollama
chromadb
yfinance
pandas
feedparser
beautifulsoup4
requests
pydub
pydantic-settings
python-multipart 
faster-whisper
apscheduler